fetch_time/pyproject.toml
[project]
name = "fetch-time"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.0.17",
    "prompt-toolkit>=3.0.51",
    "pyyaml>=6.0.2",
]
